+++ tests/testthat/helper-fixtures.R
@@ -1,0 +1,5 @@
+fixture <- function() system.file("extdata", "model.json", package = "foceexecutorr")
+fixture_v2 <- function() {
+  system.file("extdata", "model-run-fbd67a28-06a8-40b4-9cb1-e14e6cc49ff3.json",
+              package = "foceexecutorr")
+}